refactor(CrossAppLogin): Update ExternalAccount prior to 1st public release
Use a Long for the id, to not have to break compatibility if it ever grows.
Reorder properties, and change the proto numbers to be ordered while we can.

+++ CrossAppLoginUI/src/main/kotlin/com/infomaniak/core/crossloginui/views/CrossLoginListAccountsView.kt
@@ -46,7 +46,7 @@ class CrossLoginListAccountsView @JvmOverloads constructor(
 ) : AbstractComposeView(context, attrs, defStyleAttr) {
 
     private val accounts = mutableStateListOf<ExternalAccount>()
-    private val skippedIds = mutableStateSetOf<Int>()
+    private val skippedIds = mutableStateSetOf<Long>()
 
     private var onAnotherAccountClickedListener: (() -> Unit)? = null
     private var onSaveClicked: SaveListener? = null
@@ -118,7 +118,7 @@ class CrossLoginListAccountsView @JvmOverloads constructor(
         }
     }
 
-    fun setSkippedIds(items: Set<Int>) {
+    fun setSkippedIds(items: Set<Long>) {
         skippedIds.apply {
             clear()
             addAll(items)
@@ -134,6 +134,6 @@ class CrossLoginListAccountsView @JvmOverloads constructor(
     }
 
     fun interface SaveListener {
-        fun onSaveClicked(skippedAccountIds: Set<Int>)
+        fun onSaveClicked(skippedAccountIds: Set<Long>)
     }
 }
